requirements.txt
discord.py>=2.3.2
openai>=1.40.0
httpx>=0.27.0
pydantic>=2.7.0
python-dotenv>=1.0.1
PyYAML>=6.0.1
plexapi>=4.15.9
tenacity>=8.3.0
orjson>=3.10.7
rich>=13.7.1
typer>=0.12.3

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
pytest-cov>=4.1.0

